[#173362906] replace identification with confirm alert for bonus activation
