+++ locales/it/index.yml
@@ -1336,6 +1336,11 @@ bonus:
         eligibilityExpired:
           title: Eligibilità scaduta
           body: Il periodo in cui era possibile attivare il bonus è scaduto. È necessario iniziare una nuova procedura.
+        confirm:
+          title: Attenzione
+          body: Continuando nel processo, non sarà più possibile annullare la richiesta.
+          cancel: Torna alla richiesta
+          confirm: Continua
         completed:
           title: Congratulazioni!
           description: Il tuo bonus è attivo
